requirements.txt
discord.py
pytz
python-dateutil